refactor: old test の移植
